Implement ImageListViewSelectedItemCollection IList accessors
This implements previously-NotSupported IList accessors in order to fix issues where methods accepting an IEnumerable will try to optimize if a more useful interface like IList exists. For example, Linq's FirstOrDefault will fail because it can't use the IList.this[int index] accessor. And passing SelectedItems to a List constructor will fail because IList.CopyTo is not supported.

The obsolete notice on IList.IndexOf has been removed, since ImageListViewItem.Index does not return the same value as the index in the selected items collection.